requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
boto3==1.16.56
botocore==1.19.56
discord.py==1.5.1
pyparsing==2.4.7
python-dotenv==0.15.0